Implement an API for transcript based on Shake256 hash function
