+++ labrador/Cargo.toml
@@ -11,6 +11,7 @@ rand = {version = "0.9", default-features = false, features = ["thread_rng", "un
 rustfft = "6.2.0"
 thiserror = {version = "1.0", default-features = false }
 blake2 = "0.11.0-pre.5"
+sha3 = "0.10.8"
 [features]
 skip-slow-tests = []
