+++ labrador/src/poseidon/mod.rs
@@ -3,6 +3,7 @@
 //! This crate provides a **modular implementation of the
 //! Poseidon hash function, organised into permutation, sponge, and transcript sub‑modules.
 
-pub mod permutation;
-pub mod sponge;
+// pub mod permutation;
+// pub mod sponge;
 pub mod transcript;
+pub mod shake_sponge;
